+++ CHANGELOG.md
@@ -9,6 +9,7 @@ KEY CHANGES:
 - [ENH] DXGI: Expose types from windows.ui.xaml.media.dxinterop.h
 - [FIX] WIC: Crash with IWICBitmapLock.GetDataPointer, correctly handle different types
 - [FIX] Direct3D11: Fix OMSetRenderTargetsAndUnorderedAccessViews overloads and add KeepRenderTargetsAndDepthStencil and KeepUnorderedAccessViews support.
+- [ENH] DXGI: Improvements in API usage for IDXGIObject and IDXGIDeviceSubObject.
 
 -----------------------------------------------
 Release:     1.6.0 (April 2020)